Bug 1001326 - Split out GaiaApp/GaiaApps python classes into their own files
