+++ tests/python/gaia-ui-tests/gaiatest/gaia_test.py
@@ -18,108 +18,7 @@ from marionette.errors import InvalidResponseException
 from marionette.wait import Wait
 
 from file_manager import GaiaDeviceFileManager, GaiaLocalFileManager
-
-
-class GaiaApp(object):
-
-    def __init__(self, origin=None, name=None, frame=None, src=None):
-        self.frame = frame
-        self.frame_id = frame
-        self.src = src
-        self.name = name
-        self.origin = origin
-
-    def __eq__(self, other):
-        return self.__dict__ == other.__dict__
-
-
-class GaiaApps(object):
-
-    def __init__(self, marionette):
-        self.marionette = marionette
-        js = os.path.abspath(os.path.join(__file__, os.path.pardir, 'atoms', "gaia_apps.js"))
-        self.marionette.import_script(js)
-
-    def get_permission(self, app_name, permission_name):
-        self.marionette.switch_to_frame()
-        return self.marionette.execute_async_script("return GaiaApps.getPermission('%s', '%s')" % (app_name, permission_name))
-
-    def set_permission(self, app_name, permission_name, value):
-        self.marionette.switch_to_frame()
-        return self.marionette.execute_async_script("return GaiaApps.setPermission('%s', '%s', '%s')" %
-                                                    (app_name, permission_name, value))
-
-    def launch(self, name, switch_to_frame=True, launch_timeout=None):
-        self.marionette.switch_to_frame()
-        result = self.marionette.execute_async_script("GaiaApps.launchWithName('%s')" % name, script_timeout=launch_timeout)
-        assert result, "Failed to launch app with name '%s'" % name
-        app = GaiaApp(frame=result.get('frame'),
-                      src=result.get('src'),
-                      name=result.get('name'),
-                      origin=result.get('origin'))
-        if app.frame_id is None:
-            raise Exception("App failed to launch; there is no app frame")
-        if switch_to_frame:
-            self.marionette.switch_to_frame(app.frame_id)
-        return app
-
-    @property
-    def displayed_app(self):
-        self.marionette.switch_to_frame()
-        result = self.marionette.execute_script('return GaiaApps.getDisplayedApp();')
-        return GaiaApp(frame=result.get('frame'),
-                       src=result.get('src'),
-                       name=result.get('name'),
-                       origin=result.get('origin'))
-
-    def switch_to_displayed_app(self):
-        self.marionette.switch_to_default_content()
-        self.marionette.switch_to_frame(self.displayed_app.frame)
-
-    def is_app_installed(self, app_name):
-        self.marionette.switch_to_frame()
-        return self.marionette.execute_async_script("GaiaApps.locateWithName('%s')" % app_name)
-
-    def uninstall(self, name):
-        self.marionette.switch_to_frame()
-        self.marionette.execute_async_script("GaiaApps.uninstallWithName('%s')" % name)
-
-    def kill(self, app):
-        self.marionette.switch_to_frame()
-        result = self.marionette.execute_async_script("GaiaApps.kill('%s');" % app.origin)
-        assert result, "Failed to kill app with name '%s'" % app.name
-
-    def kill_all(self):
-        self.marionette.switch_to_frame()
-        self.marionette.execute_async_script("GaiaApps.killAll()")
-
-    @property
-    def installed_apps(self):
-        apps = self.marionette.execute_async_script(
-            'return GaiaApps.getInstalledApps();')
-        result = []
-        for app in [a for a in apps if not a['manifest'].get('role')]:
-            entry_points = app['manifest'].get('entry_points')
-            if entry_points:
-                for ep in entry_points.values():
-                    result.append(GaiaApp(
-                        origin=app['origin'],
-                        name=ep['name']))
-            else:
-                result.append(GaiaApp(
-                    origin=app['origin'],
-                    name=app['manifest']['name']))
-        return result
-
-    @property
-    def running_apps(self):
-        self.marionette.switch_to_frame()
-        apps = self.marionette.execute_script(
-            'return GaiaApps.getRunningApps();')
-        result = []
-        for app in [a[1] for a in apps.items()]:
-            result.append(GaiaApp(origin=app['origin'], name=app['name']))
-        return result
+from gaia_apps import GaiaApps
 
 
 class GaiaData(object):
